requirements.txt
Click==8.1.3
Flask==2.1.2
flask_jwt_extended==4.4.0
pyjwt==2.4.0
jsonschema==4.5.1
requests==2.27.1
netaddr==0.8.0
pyopenssl==22.0.0
psutil==5.9.1
pylxd==2.3.1
terminado==0.15.0
tornado==6.1
tornado-xstatic
XStatic-term.js==0.0.7.0
bcrypt==3.2.2
pyyaml==6.0
